[parser] Updated parser control flow
